+++ src/powershell/data/CommunityCustomApps.ps1
@@ -29,6 +29,6 @@ $global:CommunityCustomApps = [ordered]@{
         AssetPath   = 'librespot'
         RequiredFiles = @('manifest.json', 'index.js', 'style.css', 'librespot-engine.js', 'LICENSE', 'THIRD_PARTY_NOTICES.md')
         CompanionExtension = 'librespot-engine.js'
-        SHA256      = '36e5a478cb122c79a85a4794d8d54f7400ce3e250c75a097236a56babe5a0d9d'
+        SHA256      = '66ece851582abdb8d0ec6d4860f528a9a99e8567364ab26d6105a2c2e143ad2a'
     }
 }